+++ src/main/java/com/nhnacademy/member_server/controller/MemberController.java
@@ -6,8 +6,8 @@ import com.nhnacademy.member_server.dto.response.member.SimpleMemberResponse;
 import com.nhnacademy.member_server.entity.MemberPrincipal;
 import com.nhnacademy.member_server.entity.Role;
 import com.nhnacademy.member_server.global.jwt.WebUtils;
-import com.nhnacademy.member_server.service.AuthService;
-import com.nhnacademy.member_server.service.MemberService;
+import com.nhnacademy.member_server.service.member.AuthService;
+import com.nhnacademy.member_server.service.member.MemberService;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
