+++ src/main/java/com/nhnacademy/member_server/service/impl/member/AddressServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.member_server.service.impl;
+package com.nhnacademy.member_server.service.impl.member;
 
 import com.nhnacademy.member_server.dto.request.member.AddressRequest;
 import com.nhnacademy.member_server.dto.response.member.AddressListResponse;
@@ -7,7 +7,7 @@ import com.nhnacademy.member_server.entity.Address;
 import com.nhnacademy.member_server.entity.Member;
 import com.nhnacademy.member_server.repository.AddressRepository;
 import com.nhnacademy.member_server.repository.MemberRepository;
-import com.nhnacademy.member_server.service.AddressService;
+import com.nhnacademy.member_server.service.member.AddressService;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
