+++ src/main/java/com/nhnacademy/member_server/service/member/MemberService.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.member_server.service;
+package com.nhnacademy.member_server.service.member;
 
 import com.nhnacademy.member_server.dto.request.member.MemberUpdateRequest;
 import com.nhnacademy.member_server.dto.response.member.MemberResponse;
