+++ src/test/java/com/nhnacademy/member_server/controller/MemberControllerTest.java
@@ -5,8 +5,8 @@ import com.nhnacademy.member_server.dto.request.member.MemberUpdateRequest;
 import com.nhnacademy.member_server.dto.response.member.MemberResponse;
 import com.nhnacademy.member_server.entity.Gender;
 import com.nhnacademy.member_server.entity.MemberPrincipal;
-import com.nhnacademy.member_server.service.AuthService;
-import com.nhnacademy.member_server.service.MemberService;
+import com.nhnacademy.member_server.service.member.AuthService;
+import com.nhnacademy.member_server.service.member.MemberService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
